% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianfingerprintr-package.R
\docType{package}
\name{bayesianfingerprintr-package}
\alias{bayesianfingerprintr}
\alias{bayesianfingerprintr-package}
\title{bayesianfingerprintr: Inferring the Age and Sex of Ancient Potters from Fingerprint Ridge Densities}
\description{
The density of epidermal ridges in a fingerprint varies predictably by age and sex. Archaeologists are therefore interested in using recovered fingerprints to learn about the ancient people who produced them. Recent studies focus on estimating the age and sex of individuals by measuring their fingerprints with one of two similar metrics: mean ridge breadth (MRB) or ridge density (RD). Yet these attempts face several critical problems: expected values for adult females and adolescent males are inherently indistinguishable, and inter-assemblage variation caused by biological and technological differences cannot be easily estimated. Each of these factors greatly decreases the accuracy of predictions based on individual prints, and together they condemn this strategy to relative uselessness. However, information in fingerprints from across an assemblage can be pooled to generate a more accurate depiction of potter demographics. We present a new approach to epidermal ridge density analysis using Bayesian mixture models with the following key benefits: 1) Age and sex are estimated more accurately than existing methods by incorporating a data-driven understanding of how demographics and ridge density covary. 2) Uncertainty in demographic estimates is automatically quantified and included in output. 3) The Bayesian framework can be easily adapted to fit the unique needs of different researchers.
}
\author{
\strong{Maintainer}: Andrew Burchill \email{andrew.burchill@asu.edu} (\href{https://orcid.org/0000-0002-4197-6689}{ORCID})

}
\keyword{internal}
