% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModel.R
\name{make_modelstring}
\alias{make_modelstring}
\title{Create the JAGS model file}
\usage{
make_modelstring(
  model.opt = c("2sex.2age", "1sex.1age", "2sex.1age", "1sex.2age", "compare1v2age",
    "compare1v2ageNEXT"),
  filename = "bayesmodel.txt",
  pseudopriors = NULL
)
}
\arguments{
\item{model.opt}{a string representing which version of the model you want to run}

\item{filename}{the name of the text document that the model will be written to}

\item{pseudopriors}{when running a "super model," this is where you pass the posterior distributions of the two models to then be used as pseudopriors}
}
\value{
a list of the nodes (parameters) that should be monitored during the JAGS MCMC run
}
\description{
This writes the desired fingerprint mixture model into a separate file to be read into JAGS. It also returns the nodes that need to be monitored during the MCMC run.
}
\examples{
print("hello")

}
\author{
Andrew Burchill, \email{andrew.burchill@asu.edu}
}
