% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalingFactor.R
\name{scaling_factor}
\alias{scaling_factor}
\title{Align MRB assemblages to the modern reference dataset}
\usage{
scaling_factor(mrbs, x = 0.95, type = "all", adult_age = 15)
}
\arguments{
\item{mrbs}{the vector or column of your collected MRB measurements}

\item{x}{what percentile should serve as the alignment point? (defaults to 95th percentile)}

\item{type}{are you aligning your data to a subset of the reference data? Options are currently \code{"all","males","females","adult males","adult females","adults"}. We highly suggest sticking with the default unless there is strong evidence otherwise.}

\item{adult_age}{what age do you consider to be the threshold for adulthood? (only used when scaling to a subset of the reference data)}
}
\value{
the scaling factor that your data should be multiplied by in order to be analagous with modern MRB data
}
\description{
Interassemblage variation makes comparing mean ridge breadth (MRB) measurements
difficult. This function scales a vector of MRB values to the modern
Kralik and Novotny (2003) reference data set.
}
\examples{
fake.data <-
data.frame(
 id.number = 1:100,
  mrb = c(rgamma(50,64,160), rgamma(50,2500,5000)),
  location = sample(c("palace", "domestic"), 100, replace = TRUE)
)

fake.data$mrb <- fake.data$mrb*scaling_factor(fake.data$mrb)

}
